fixed dopebox override url
