+++ app/src/main/java/com/lagradost/cloudstream3/movieproviders/BflixProvider.kt
@@ -7,16 +7,7 @@ import com.lagradost.cloudstream3.utils.ExtractorLink
 import com.lagradost.cloudstream3.utils.loadExtractor
 import org.jsoup.Jsoup
 
-class FmoviesToProvider : BflixProvider() {
-    override var mainUrl = "https://fmovies.to"
-    override var name = "Fmovies.to"
-}
-class SflixProProvider : BflixProvider() {
-    override var mainUrl = "https://sflix.pro"
-    override var name = "Sflix.pro"
-}
-
-open class BflixProvider() : MainAPI() {
+open class BflixProvider : MainAPI() {
     override var mainUrl = "https://bflix.to"
     override var name = "Bflix"
     override val hasMainPage = true
