+++ app/src/main/java/com/lagradost/cloudstream3/movieproviders/DopeboxProvider.kt
@@ -1,0 +1,6 @@
+package com.lagradost.cloudstream3.movieproviders
+
+class DopeboxProvider : SflixProvider() {
+    override var mainUrl = "https://dopebox.to"
+    override var name = "Dopebox"
+}
