+++ app/src/main/java/com/lagradost/cloudstream3/movieproviders/FmoviesToProvider.kt
@@ -1,0 +1,6 @@
+package com.lagradost.cloudstream3.movieproviders
+
+class FmoviesToProvider : BflixProvider() {
+    override var mainUrl = "https://fmovies.to"
+    override var name = "Fmovies.to"
+}
