+++ app/src/main/java/com/lagradost/cloudstream3/movieproviders/SflixProProvider.kt
@@ -1,0 +1,6 @@
+package com.lagradost.cloudstream3.movieproviders
+
+class SflixProProvider : BflixProvider() {
+    override var mainUrl = "https://sflix.pro"
+    override var name = "Sflix.pro"
+}
