+++ app/src/main/java/com/lagradost/cloudstream3/movieproviders/SolarmovieProvider.kt
@@ -1,0 +1,6 @@
+package com.lagradost.cloudstream3.movieproviders
+
+class SolarmovieProvider : SflixProvider() {
+    override var mainUrl = "https://solarmovie.pe"
+    override var name = "Solarmovie"
+}
